requirements.txt
matplotlib
numpy
Pillow
PyYAML
scikit-learn
torch
torchmetrics
torchvision
pytest
